+++ package/data/db/configFilter.json.js
@@ -1,0 +1,1 @@
+jsonjs.data['configFilter'] = [["entity","field","id","is_active_default","name","value"],["dataset","type","closed_data",true,"Closed Data","closed_data"],["dataset","type","open_data",true,"Open Data","open_data"]]